+++ aspen/exceptions.py
@@ -12,7 +12,3 @@ class LoadError(Exception):
     """
     # Define this here to avoid import issues when json doesn't exist.
 
-
-class CRLFInjection(Exception):
-    def __str__(self):
-        return "Possible CRLF injection detected."
