+++ tests/test_mappings.py
@@ -9,7 +9,7 @@ from aspen import Response
 
 from aspen.http.mapping import Mapping, CaseInsensitiveMapping
 
-from aspen.http.baseheaders import BaseHeaders
+from aspen.http.response import BaseHeaders
 from aspen.http.request import Querystring
